move operation marker interfaces from spring-server to types
Those generic marker interfaces can be used by different server implementations and are also needed by plugins to be able to generate SDL.
