+++ examples/federation/base-app/src/main/kotlin/com/expediagroup/graphql/examples/query/SimpleQuery.kt
@@ -17,7 +17,7 @@
 package com.expediagroup.graphql.examples.query
 
 import com.expediagroup.graphql.annotations.GraphQLDescription
-import com.expediagroup.graphql.spring.operations.Query
+import com.expediagroup.graphql.server.operations.Query
 import org.springframework.stereotype.Component
 
 @Suppress("FunctionOnlyReturningConstant")
