+++ examples/server/spring-server/src/main/kotlin/com/expediagroup/graphql/examples/query/AsyncQuery.kt
@@ -17,7 +17,7 @@
 package com.expediagroup.graphql.examples.query
 
 import com.expediagroup.graphql.annotations.GraphQLDescription
-import com.expediagroup.graphql.spring.operations.Query
+import com.expediagroup.graphql.server.operations.Query
 import kotlinx.coroutines.coroutineScope
 import kotlinx.coroutines.time.delay
 import org.springframework.stereotype.Component
