+++ examples/server/spring-server/src/main/kotlin/com/expediagroup/graphql/examples/query/DataAndErrorsQuery.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright 2019 Expedia, Inc
+ * Copyright 2020 Expedia, Inc
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -16,8 +16,8 @@
 
 package com.expediagroup.graphql.examples.query
 
+import com.expediagroup.graphql.server.operations.Query
 import com.expediagroup.graphql.spring.exception.SimpleKotlinGraphQLError
-import com.expediagroup.graphql.spring.operations.Query
 import graphql.execution.DataFetcherResult
 import graphql.execution.ResultPath
 import graphql.language.SourceLocation
